+++ spring-integration-core/src/main/java/org/springframework/integration/json/ObjectToJsonTransformer.java
@@ -18,6 +18,9 @@ package org.springframework.integration.json;
 import org.springframework.integration.Message;
 import org.springframework.integration.MessageHeaders;
 import org.springframework.integration.support.MessageBuilder;
+import org.springframework.integration.support.json.JacksonJsonObjectMapper;
+import org.springframework.integration.support.json.JacksonJsonObjectMapperProvider;
+import org.springframework.integration.support.json.JsonObjectMapper;
 import org.springframework.integration.transformer.AbstractTransformer;
 import org.springframework.util.Assert;
 import org.springframework.util.ClassUtils;
