+++ spring-integration-core/src/main/java/org/springframework/integration/support/json/Jackson2JsonMessageParser.java
@@ -14,31 +14,33 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.json;
+package org.springframework.integration.support.json;
+
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import org.codehaus.jackson.JsonFactory;
-import org.codehaus.jackson.JsonParser;
-import org.codehaus.jackson.JsonToken;
-import org.codehaus.jackson.map.JsonMappingException;
 import org.springframework.integration.Message;
 import org.springframework.integration.support.MessageBuilder;
 import org.springframework.util.Assert;
 
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonToken;
+import com.fasterxml.jackson.databind.JsonMappingException;
+
 /**
  * {@link JsonInboundMessageMapper.JsonMessageParser} implementation that parses JSON messages
  * and builds a {@link Message} with the specified payload type from provided {@link JsonInboundMessageMapper}.
- * Uses Jackson JSON-processor (@link http://jackson.codehaus.org).
+ * Uses Jackson 2 JSON-processor (@link https://github.com/FasterXML).
  *
  * @author Artem Bilan
  * @since 3.0
  */
-class JacksonJsonMessageParser extends AbstractJacksonJsonMessageParser<JsonParser> {
+class Jackson2JsonMessageParser extends AbstractJacksonJsonMessageParser<JsonParser> {
 
-	JacksonJsonMessageParser() {
-		super(new JacksonJsonObjectMapper());
+	public Jackson2JsonMessageParser() {
+		super(new Jackson2JsonObjectMapper());
 	}
 
 	@Override
