+++ spring-integration-core/src/main/java/org/springframework/integration/support/json/JsonObjectMapperAdapter.java
@@ -14,33 +14,43 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.json;
+package org.springframework.integration.support.json;
 
 import java.io.Reader;
 import java.io.Writer;
 import java.lang.reflect.Type;
 
 /**
- * Strategy interface to convert an Object to/from the JSON representation.
+ * Simple {@linkplain JsonObjectMapper} adapter implementation, if there is no need
+ * to provide entire operations implementation.
  *
  * @author Artem Bilan
  * @since 3.0
- *
- * @see JsonToObjectTransformer
- * @see ObjectToJsonTransformer
- * @see JacksonJsonObjectMapper
- * @see Jackson2JsonObjectMapper
  */
-public interface JsonObjectMapper<P> {
+public abstract class JsonObjectMapperAdapter<P> implements JsonObjectMapper<P> {
 
-	String toJson(Object value) throws Exception;
+	@Override
+	public String toJson(Object value) throws Exception {
+		return null;
+	}
 
-	void toJson(Object value, Writer writer) throws Exception;
+	@Override
+	public void toJson(Object value, Writer writer) throws Exception {
+	}
 
-	<T> T fromJson(String json, Class<T> valueType) throws Exception;
+	@Override
+	public <T> T fromJson(String json, Class<T> valueType) throws Exception {
+		return null;
+	}
 
-	<T> T fromJson(Reader json, Class<T> valueType) throws Exception;
+	@Override
+	public <T> T fromJson(Reader json, Class<T> valueType) throws Exception {
+		return null;
+	}
 
-	<T> T fromJson(P parser, Type valueType) throws Exception;
+	@Override
+	public <T> T fromJson(P parser, Type valueType) throws Exception {
+		return null;
+	}
 
 }
