+++ spring-integration-core/src/main/java/org/springframework/integration/support/json/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides classes supporting json.
+ */
+package org.springframework.integration.support.json;
